+++ packages/atomic/src/components/atomic-query-error/atomic-query-error.tsx
@@ -6,16 +6,29 @@ import {
   InitializableComponent,
   InitializeBindings,
 } from '../../utils/initialization-utils';
+import {Button} from '../common/button';
+import CannotAccess from '../../images/cannot-access.svg';
+import SomethingWrong from '../../images/something-wrong.svg';
+import SearchInactive from '../../images/search-inactive.svg';
+import NoConnection from '../../images/no-connection.svg';
+import Indexing from '../../images/indexing.svg';
 
-// TODO: Get type from Headless once the API provides more streamlined errors.
 const disconnectedException = 'Disconnected';
 const noEndpointsException = 'NoEndpointsException';
 const invalidTokenException = 'InvalidTokenException';
 const organizationIsPausedException = 'OrganizationIsPausedException';
 
+interface QueryErrorDetails {
+  icon: string;
+  title: string;
+  description: string;
+  link?: string;
+}
+
 /**
  * The `atomic-query-error` component handles fatal errors when performing a query on the index or Search API. When the error is known, it displays a link to relevant documentation link for debugging purposes. When the error is unknown, it displays a small text area with the JSON content of the error.
  *
+ * @part icon - The svg related to the error.
  * @part title - The title of the error.
  * @part description - A description of the error.
  * @part doc-link - A link to the relevant documentation.
@@ -31,21 +44,6 @@ export class AtomicQueryError implements InitializableComponent {
   @InitializeBindings() public bindings!: Bindings;
   public queryError!: QueryError;
 
-  private strings = {
-    disconnectedTitle: () => this.bindings.i18n.t('disconnected'),
-    disconnectedDesc: () => this.bindings.i18n.t('check-your-connection'),
-    noEndpointsTitle: () => this.bindings.i18n.t('no-endpoints'),
-    noEndpointsDesc: () => this.bindings.i18n.t('add-sources'),
-    invalidTokenTitle: () => this.bindings.i18n.t('cannot-access'),
-    invalidTokenDesc: () => this.bindings.i18n.t('invalid-token'),
-    genericErrorTitle: () => this.bindings.i18n.t('something-went-wrong'),
-    genericErrorDesc: () => this.bindings.i18n.t('if-problem-persists'),
-    helpLink: () => this.bindings.i18n.t('coveo-online-help'),
-    moreInfo: () => this.bindings.i18n.t('more-info'),
-    organizationIsPaused: () => this.bindings.i18n.t('organization-is-paused'),
-    organizationWillResume: () =>
-      this.bindings.i18n.t('organization-will-resume'),
-  };
   @BindStateToController('queryError')
   @State()
   private queryErrorState!: QueryErrorState;
@@ -59,66 +57,73 @@ export class AtomicQueryError implements InitializableComponent {
   private renderShowMoreInfo() {
     if (!this.showMoreInfo) {
       return (
-        <button
+        <Button
           part="more-info-btn"
-          class="text-primary"
+          style="primary"
+          class="p-3 mt-8"
           onClick={() => (this.showMoreInfo = true)}
-        >
-          {this.strings.moreInfo()}
-        </button>
+          text={this.bindings.i18n.t('more-info')}
+        ></Button>
       );
     }
 
     return (
       <pre
         part="error-info"
-        class="text-left border border-neutral bg-neutral-light p-3 rounded my-4 whitespace-pre-wrap"
+        class="text-left border border-neutral bg-neutral-light p-3 rounded mt-8 whitespace-pre-wrap"
       >
         <code>{JSON.stringify(this.queryErrorState.error, null, 2)}</code>
       </pre>
     );
   }
 
-  private get title() {
-    switch (this.queryErrorState.error!.type) {
-      case disconnectedException:
-        return this.strings.disconnectedTitle();
-      case noEndpointsException:
-        return this.strings.noEndpointsTitle();
-      case invalidTokenException:
-        return this.strings.invalidTokenTitle();
-      case organizationIsPausedException:
-        return this.strings.organizationWillResume();
-      default:
-        return this.strings.genericErrorTitle();
-    }
+  private get org() {
+    return this.bindings.engine.state.configuration.organizationId;
   }
 
-  private get description() {
-    switch (this.queryErrorState.error!.type) {
-      case disconnectedException:
-        return this.strings.disconnectedDesc();
-      case noEndpointsException:
-        return this.strings.noEndpointsDesc();
-      case invalidTokenException:
-        return this.strings.invalidTokenDesc();
-      case organizationIsPausedException:
-        return this.strings.organizationIsPaused();
-      default:
-        return this.strings.genericErrorDesc();
-    }
+  private get errorType() {
+    return this.queryErrorState.error!.type;
   }
 
-  private get link() {
-    switch (this.queryErrorState.error!.type) {
+  private get details(): QueryErrorDetails {
+    switch (this.errorType) {
+      case disconnectedException:
+        return {
+          icon: NoConnection,
+          title: this.bindings.i18n.t('disconnected'),
+          description: this.bindings.i18n.t('check-your-connection'),
+        };
       case noEndpointsException:
-        return 'https://docs.coveo.com/'; // TODO: update link
+        return {
+          icon: Indexing,
+          title: this.bindings.i18n.t('no-endpoints', {org: this.org}),
+          description: this.bindings.i18n.t('add-sources'),
+          link: 'https://docs.coveo.com/', // TODO: KIT-1061 update link
+        };
       case invalidTokenException:
-        return 'https://docs.coveo.com/'; // TODO: update link
+        return {
+          icon: CannotAccess,
+          title: this.bindings.i18n.t('cannot-access', {org: this.org}),
+          description: this.bindings.i18n.t('invalid-token'),
+          link: 'https://docs.coveo.com/', // TODO: KIT-1061 update link
+        };
       case organizationIsPausedException:
-        return 'https://docs.coveo.com/l6af0467';
+        return {
+          icon: SearchInactive,
+          title: this.bindings.i18n.t('organization-is-paused', {
+            org: this.org,
+          }),
+          description: this.bindings.i18n.t('organization-will-resume', {
+            org: this.org,
+          }),
+          link: 'https://docs.coveo.com/l6af0467',
+        };
       default:
-        return null;
+        return {
+          icon: SomethingWrong,
+          title: this.bindings.i18n.t('something-went-wrong'),
+          description: this.bindings.i18n.t('if-problem-persists'),
+        };
     }
   }
 
@@ -129,19 +134,24 @@ export class AtomicQueryError implements InitializableComponent {
 
     return (
       <div class="text-center">
-        <h3 part="title" class="text-2xl text-primary my-4">
-          {this.title}
+        <atomic-icon
+          part="icon"
+          icon={this.details.icon}
+          class="w-1/2 mt-8"
+        ></atomic-icon>
+        <h3 part="title" class="text-2xl text-on-background mt-8">
+          {this.details.title}
         </h3>
-        <p part="description" class="text-xl text-primary my-4">
-          {this.description}
+        <p part="description" class="text-lg text-neutral-dark mt-2.5">
+          {this.details.description}
         </p>
-        {this.link ? (
+        {this.details.link ? (
           <a
-            href={this.link}
+            href={this.details.link}
             part="doc-link"
-            class="text-primary text-xl hover:underline"
+            class="btn-primary p-3 mt-10 inline-block"
           >
-            {this.strings.helpLink()}
+            {this.bindings.i18n.t('coveo-online-help')}
           </a>
         ) : (
           this.renderShowMoreInfo()