+++ packages/headless/src/features/fields/fields-slice.test.ts
@@ -1,6 +1,18 @@
 import {fieldsReducer} from './fields-slice';
-import {registerFieldsToInclude} from './fields-actions';
+import {
+  disableFetchAllFields,
+  enableFetchAllFields,
+  fetchFieldsDescription,
+  registerFieldsToInclude,
+} from './fields-actions';
 import {FieldsState, getFieldsInitialState} from './fields-state';
+import {
+  buildMockSearchAppEngine,
+  createMockState,
+  MockSearchEngine,
+} from '../../test';
+import {PlatformClient} from '../../api/platform-client';
+import {buildMockFieldDescription} from '../../test/mock-field-description';
 
 describe('fields slice', () => {
   it('should have initial state', () => {
@@ -26,5 +38,56 @@ describe('fields slice', () => {
   it('should not have duplicate fields', () => {
     const action = registerFieldsToInclude(['author', 'language']);
     expect(fieldsReducer(undefined, action)).toEqual(getFieldsInitialState());
+  });
+
+  it('should fetch all fields if requested', () => {
+    expect(
+      fieldsReducer(getFieldsInitialState(), enableFetchAllFields())
+        .fetchAllFields
+    ).toBe(true);
+  });
+
+  it('should not fetch all fields if requested', () => {
+    expect(
+      fieldsReducer(getFieldsInitialState(), disableFetchAllFields())
+        .fetchAllFields
+    ).toBe(false);
+  });
+
+  describe('when fetching fields description', () => {
+    let e: MockSearchEngine;
+
+    beforeEach(() => {
+      e = buildMockSearchAppEngine({state: createMockState()});
+      PlatformClient.call = jest.fn().mockImplementation(() => {
+        const body = JSON.stringify([
+          buildMockFieldDescription(),
+          buildMockFieldDescription(),
+        ]);
+        const response = new Response(body);
+
+        return Promise.resolve(response);
+      });
+    });
+
+    it('calls platform client to get fields description', async () => {
+      await e.dispatch(fetchFieldsDescription());
+      expect(PlatformClient.call).toHaveBeenCalledWith(
+        expect.objectContaining({
+          url: expect.stringContaining('/fields'),
+          method: 'GET',
+        })
+      );
+    });
+
+    it('should add fields description to the state', () => {
+      const action = fetchFieldsDescription.fulfilled(
+        [buildMockFieldDescription(), buildMockFieldDescription()],
+        ''
+      );
+      expect(
+        fieldsReducer(getFieldsInitialState(), action).fieldsDescription.length
+      ).toBe(2);
+    });
   });
 });