+++ packages/quantic/cypress/integration/console-expectations.ts
@@ -1,0 +1,6 @@
+import {ConsoleSelectors} from './console-selectors';
+
+export const ConsoleExpectations = {
+  error: (hasError: boolean) =>
+    ConsoleSelectors.error().should(hasError ? 'be.called' : 'not.be.called'),
+};