+++ packages/quantic/force-app/main/default/lwc/quanticUtils/quanticUtils.js
@@ -213,3 +213,16 @@ export class TimeSpan {
     return this.getHHMMSS().replace(/^0+/, '');
   }
 }
+
+/**
+ * Converts a date string from the Coveo Search API format to the ISO-8601 format.
+ * Replace `/` characters in date string with `-`.
+ * Replace `@` characters in date string with `T`.
+ * @param {string} dateString 
+ * @returns {string}
+ */
+export function fromSearchApiDate(dateString) {
+  return dateString
+    .replaceAll('/', '-')
+    .replaceAll('@', 'T');
+}